requirements.txt
feedgenerator>=1.6
python-twitter>=1.0
